feat: return value from http server
This introduces some quick hacks such that the async nature of axum
works with our LSM store. A lot of these methods were originally taking
mutable references which become very difficult to handle once we're in
an async networking context.

This commit simply throws arc/mutexes at the issues for now so that the
store works over HTTP. Cleanup will be required afterwards.