feat: correct buffer flushing within wal for restoration
The buffer was not being cleared which greatly polluted the segment size
tracking and meant that the buffer size was unbounded. All data that was
written was then being rewritten again on the next write, alongside the
new data.

This bundles in the WAL restoration into another allocated buffer for
the `restore` functionality moved into the WAL.