refactor: move restore to wal (#21)
* refactor: move restore to wal

Restoring the WAL into the memtable doesn't make much sense from a
responsibility perspective. The WAL knows of the segments and should
run the restore operation. It can expose an iterator for reading
directly from itself later such that the memtable can be repopulated.

* feat: wip memtable restore

* chore: deps

* feat: correct buffer flushing within wal for restoration

The buffer was not being cleared which greatly polluted the segment size
tracking and meant that the buffer size was unbounded. All data that was
written was then being rewritten again on the next write, alongside the
new data.

This bundles in the WAL restoration into another allocated buffer for
the `restore` functionality moved into the WAL.

* fix: read file not log dir

* feat: add restore assertions

* docs: clarify append_batch invariant

* feat: skip invalid utf8 entries in wal